Create bday_paradox.cpp
A simulator that explores the birthday paradox, which states that if you were to put 100 people in a room, you're guaranteed to have at least one pair of people who were born on the same day! (Not accounting for birth year). This program uses random numbers, structs, and operator overwriting.